fix: 增加性能优化
