+++ src/packages/swiper/index.ts
@@ -2,7 +2,7 @@ import React from 'react'
 import { Swiper, SwiperProps } from './swiper'
 import { SwiperRef } from './types'
 import { SwiperItem } from '../swiperitem/swiperitem'
-import { Focus } from './focus'
+import { SwiperFocus } from './swiper-focus'
 
 export type { SwiperProps } from './swiper'
 export type { SwiperIndicator, SwiperRef, SwiperDirections } from './types'
@@ -12,9 +12,9 @@ type CompoundedComponent = React.ForwardRefExoticComponent<
     React.RefAttributes<SwiperRef>
 > & {
   Item: typeof SwiperItem
-  Focus: typeof Focus
+  Focus: typeof SwiperFocus
 }
 const InnerSwiper = Swiper as CompoundedComponent
 InnerSwiper.Item = SwiperItem
-InnerSwiper.Focus = Focus
+InnerSwiper.Focus = SwiperFocus
 export default InnerSwiper
